refactor dpm campaign launch action
