+++ src/app/router_registry.py
@@ -114,6 +114,9 @@ from app.routers.dpm_wave_campaign_discovery import (
 from app.routers.dpm_wave_campaign_launch import (
     router as dpm_wave_campaign_launch_router,
 )
+from app.routers.dpm_wave_campaign_launch_actions import (
+    router as dpm_wave_campaign_launch_actions_router,
+)
 from app.routers.dpm_wave_campaign_lifecycle import (
     router as dpm_wave_campaign_lifecycle_router,
 )
@@ -282,6 +285,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_evidence_router)
     app.include_router(dpm_wave_ai_router)
     app.include_router(dpm_wave_campaign_launch_router)
+    app.include_router(dpm_wave_campaign_launch_actions_router)
     app.include_router(dpm_wave_campaign_lifecycle_router)
     app.include_router(dpm_wave_campaign_readiness_router)
     app.include_router(workbench_router)
